+++ src/Model/Table/PostsTable.php
@@ -31,7 +31,8 @@ use Tools\Exceptionist;
 
 /**
  * Posts model
- * @property \Cake\ORM\Association\BelongsTo $Users
+ * @property \Cake\ORM\Association\BelongsTo&\MeCms\Model\Table\PostsCategoriesTable $Categories
+ * @property \Cake\ORM\Association\BelongsTo&\MeCms\Model\Table\UsersTable $Users
  * @property \Cake\ORM\Association\BelongsToMany&\MeCms\Model\Table\TagsTable $Tags
  * @method \MeCms\Model\Entity\Post get($primaryKey, $options = [])
  * @method \MeCms\Model\Entity\Post newEntity($data = null, array $options = [])